job_backend/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-light==0.1.0.post3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-writable-nested==0.7.0
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.5.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
tzlocal==4.2
